+++ src/automigrate/constraints.clj
@@ -3,8 +3,9 @@
 
 
 (def ^:private PRIMARY-KEY-CONSTRAINT-POSTFIX "pkey")
-(def ^:private UNIQUE-INDEX-POSTFIX "key")
-(def ^:private FOREIGN-KEY-INDEX-POSTFIX "fkey")
+(def ^:private UNIQUE-CONSTRAINT-POSTFIX "key")
+(def ^:private FOREIGN-KEY-CONSTRAINT-POSTFIX "fkey")
+(def ^:private CHECK-CONSTRAINT-POSTFIX "check")
 
 
 (defn primary-key-constraint-name
@@ -16,13 +17,20 @@
 
 (defn unique-constraint-name
   [model-name field-name]
-  (->> [(name model-name) (name field-name) UNIQUE-INDEX-POSTFIX]
+  (->> [(name model-name) (name field-name) UNIQUE-CONSTRAINT-POSTFIX]
     (str/join #"-")
     (keyword)))
 
 
-(defn foreign-key-index-name
+(defn foreign-key-constraint-name
   [model-name field-name]
-  (->> [(name model-name) (name field-name) FOREIGN-KEY-INDEX-POSTFIX]
+  (->> [(name model-name) (name field-name) FOREIGN-KEY-CONSTRAINT-POSTFIX]
+    (str/join #"-")
+    (keyword)))
+
+
+(defn check-constraint-name
+  [model-name field-name]
+  (->> [(name model-name) (name field-name) CHECK-CONSTRAINT-POSTFIX]
     (str/join #"-")
     (keyword)))