+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ## 0.3.1 - UNRELEASED
 
+### Added
+
+- Add fixed name for constraints.
+- Add ability to manage column level CHECK constraints.
+
 ### Fixed
 
 - Fix validation of referenced field for foreign key constraint.